pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openadapt-grounding"
version = "0.1.0"
description = "Robust UI element localization for automation"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Richard Abrich", email = "richard@openadapt.ai"}
]
maintainers = [
    {name = "OpenAdaptAI", email = "info@openadapt.ai"}
]
keywords = ["grounding", "ui", "localization", "automation", "vlm", "vision", "ocr", "omniparser"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
]
deploy = [
    "boto3>=1.34.0",
    "paramiko>=3.4.0",
    "pydantic-settings>=2.0.0",
    "fire>=0.5.0",
]
uitars = [
    "openai>=1.0.0",
]
eval = [
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "fire>=0.5.0",
    "pydantic-settings>=2.0.0",
]
# VLM API providers
providers-anthropic = [
    "anthropic>=0.40.0",
]
providers-openai = [
    "openai>=1.0.0",
]
providers-google = [
    "google-genai>=1.0.0",
]
providers = [
    "openadapt-grounding[providers-anthropic,providers-openai,providers-google]",
]
all = [
    "openadapt-grounding[dev,deploy,uitars,eval,providers]",
]

[project.urls]
Homepage = "https://github.com/OpenAdaptAI/openadapt-grounding"
Repository = "https://github.com/OpenAdaptAI/openadapt-grounding"
Documentation = "https://github.com/OpenAdaptAI/openadapt-grounding#readme"
Issues = "https://github.com/OpenAdaptAI/openadapt-grounding/issues"
Changelog = "https://github.com/OpenAdaptAI/openadapt-grounding/releases"

[tool.hatch.build.targets.wheel]
packages = ["src/openadapt_grounding"]

[tool.ruff]
line-length = 100
